+++ src/Home.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import { Link } from './vendor/components/links';
-import DashboardPage from './components/DashboardPage';
+import DashboardPage from './utils/DashboardPage';
 import {
   AndroidIcon,
   BatteryIcon,
