+++ src/android/index.jsx
@@ -1,17 +1,17 @@
 /* eslint-disable react/no-array-index-key */
 import React, { Component } from 'react';
 import Grid from '@material-ui/core/Grid';
-import DashboardPage from '../components/DashboardPage';
-import Section from '../components/Section';
-import BugzillaGraph from '../containers/BugzillaGraph';
+import DashboardPage from '../utils/DashboardPage';
+import Section from '../utils/Section';
+import BugzillaGraph from '../bugzilla/BugzillaGraph';
 import NimbledroidSection from '../nimbledroid/NimbledroidSection';
-import PerfherderGraphContainer from '../containers/PerfherderGraphContainer';
-import RedashContainer from '../containers/RedashContainer';
+import PerfherderGraphContainer from '../utils/PerfherderGraphContainer';
+import RedashContainer from '../utils/RedashContainer';
 import { CONFIG } from '../nimbledroid/config';
 import { TP6mAggregate } from './TP6mAggregate';
 import { TimeDomain } from '../vendor/jx/domains';
 import { DetailsIcon } from '../utils/icons';
-import { showBugsUrl } from '../utils/bugzilla/query';
+import { showBugsUrl } from '../bugzilla/query';
 
 class Android extends Component {
   render() {
