+++ src/power/index.jsx
@@ -5,8 +5,8 @@ import { selectFrom } from '../vendor/vectors';
 import { COMBOS, PLATFORMS, TESTS } from './config';
 import { withNavigation } from '../vendor/components/navigation';
 import Picker from '../vendor/components/navigation/Picker';
-import DashboardPage from '../components/DashboardPage';
-import PerfherderGraphContainer from '../containers/PerfherderGraphContainer';
+import DashboardPage from '../utils/DashboardPage';
+import PerfherderGraphContainer from '../utils/PerfherderGraphContainer';
 import { timePickers } from '../utils/timePickers';
 import { TimeDomain } from '../vendor/jx/domains';
